Extract mempool logic
