+++ crates/chia-sdk-test/src/full_node_simulator.rs
@@ -32,6 +32,11 @@ use rand_chacha::ChaCha8Rng;
 
 use crate::{SimulatorError, validate_clvm_and_signature};
 
+mod chain;
+mod fast_forward;
+mod mempool;
+mod validation;
+
 const BLOCK_REWARD_AMOUNT: u64 = 2_000_000_000_000;
 const PREFARM_WALLET_INDEX: u32 = 1;
 
@@ -630,186 +635,6 @@ impl FullNodeSimulator {
         }
     }
 
-    fn insert_mempool_item(
-        &mut self,
-        tx_id: Bytes32,
-        validated: ValidatedBundle,
-    ) -> Result<(), SimulatorError> {
-        let conflicting_tx_ids = self.conflicting_mempool_tx_ids(&validated);
-        if !conflicting_tx_ids.is_empty() {
-            let conflicting_removals = conflicting_tx_ids
-                .iter()
-                .filter_map(|tx_id| self.mempool.get(tx_id))
-                .flat_map(|item| item.removals.iter().copied())
-                .collect::<IndexSet<_>>();
-            let conflicting_fees = conflicting_tx_ids
-                .iter()
-                .filter_map(|tx_id| self.mempool.get(tx_id))
-                .map(|item| item.fee)
-                .sum::<u64>();
-            let replacement = conflicting_removals
-                .iter()
-                .all(|coin_id| validated.removals.contains(coin_id))
-                && validated.fee > conflicting_fees;
-
-            if !replacement {
-                return Err(SimulatorError::Validation(ErrorCode::MempoolConflict));
-            }
-
-            for tx_id in conflicting_tx_ids {
-                self.mempool.swap_remove(&tx_id);
-            }
-        }
-
-        self.mempool.insert(tx_id, validated);
-        Ok(())
-    }
-
-    fn try_fast_forward_bundle(&self, validated: &ValidatedBundle) -> Option<SpendBundle> {
-        for (coin_id, spend) in &validated.spends {
-            if (spend.flags & ELIGIBLE_FOR_FF) == 0 {
-                continue;
-            }
-
-            for mempool_item in self.mempool.values() {
-                if !mempool_item.removals.contains(coin_id) {
-                    continue;
-                }
-
-                let Some(conflicting_spend) = mempool_item.spends.get(coin_id) else {
-                    continue;
-                };
-
-                let Some((new_coin, _)) = conflicting_spend.additions.iter().find(|(coin, _)| {
-                    coin.parent_coin_info == *coin_id
-                        && coin.puzzle_hash == spend.coin_spend.coin.puzzle_hash
-                        && coin.amount == spend.coin_spend.coin.amount
-                        && (coin.amount & 1) == 1
-                }) else {
-                    continue;
-                };
-
-                let Some(new_coin_spend) = Self::fast_forward_coin_spend(
-                    &spend.coin_spend,
-                    *new_coin,
-                    conflicting_spend.coin_spend.coin,
-                ) else {
-                    continue;
-                };
-
-                let mut coin_spends = validated.spend_bundle.coin_spends.clone();
-                let Some(existing_spend) = coin_spends
-                    .iter_mut()
-                    .find(|existing| existing.coin.coin_id() == *coin_id)
-                else {
-                    continue;
-                };
-
-                *existing_spend = new_coin_spend;
-                return Some(SpendBundle::new(
-                    coin_spends,
-                    validated.spend_bundle.aggregated_signature.clone(),
-                ));
-            }
-        }
-
-        None
-    }
-
-    fn try_fast_forward_settled_bundle(&self, spend_bundle: &SpendBundle) -> Option<SpendBundle> {
-        let constants = default_constants(SIMULATOR_GENESIS_CHALLENGE, SIMULATOR_GENESIS_CHALLENGE);
-        let conds = validate_clvm_and_signature(
-            spend_bundle,
-            11_000_000_000 / 2,
-            &constants,
-            ENABLE_KECCAK_OPS_OUTSIDE_GUARD | COMPUTE_FINGERPRINT,
-        )
-        .ok()?;
-
-        let mut coin_spends = spend_bundle.coin_spends.clone();
-        let mut rewrote_any = false;
-
-        for spend in &conds.spends {
-            let Some(record) = self.coins.get(&spend.coin_id) else {
-                continue;
-            };
-            if record.spent_block_index.is_none() {
-                continue;
-            }
-
-            let Some(index) = coin_spends
-                .iter()
-                .position(|coin_spend| coin_spend.coin.coin_id() == spend.coin_id)
-            else {
-                continue;
-            };
-
-            let mut rewritten = coin_spends[index].clone();
-            loop {
-                if (rewritten.coin.amount & 1) == 0 {
-                    break;
-                }
-                let Some(current_record) = self.coins.get(&rewritten.coin.coin_id()) else {
-                    break;
-                };
-                if current_record.spent_block_index.is_none() {
-                    break;
-                }
-                let next_coin = Coin::new(
-                    rewritten.coin.coin_id(),
-                    rewritten.coin.puzzle_hash,
-                    rewritten.coin.amount,
-                );
-                let Some(next_record) = self.coins.get(&next_coin.coin_id()) else {
-                    break;
-                };
-                let Some(next_spend) =
-                    Self::fast_forward_coin_spend(&rewritten, next_record.coin, rewritten.coin)
-                else {
-                    break;
-                };
-                rewritten = next_spend;
-                rewrote_any = true;
-            }
-
-            coin_spends[index] = rewritten;
-        }
-
-        if !rewrote_any {
-            return None;
-        }
-
-        Some(SpendBundle::new(
-            coin_spends,
-            spend_bundle.aggregated_signature.clone(),
-        ))
-    }
-
-    fn fast_forward_coin_spend(
-        coin_spend: &CoinSpend,
-        new_coin: Coin,
-        new_parent: Coin,
-    ) -> Option<CoinSpend> {
-        let mut allocator = Allocator::new_limited(500_000_000);
-        let puzzle = node_from_bytes(&mut allocator, coin_spend.puzzle_reveal.as_slice()).ok()?;
-        let solution = node_from_bytes(&mut allocator, coin_spend.solution.as_slice()).ok()?;
-        let new_solution = fast_forward_singleton(
-            &mut allocator,
-            puzzle,
-            solution,
-            &coin_spend.coin,
-            &new_coin,
-            &new_parent,
-        )
-        .ok()?;
-        let new_solution_bytes = node_to_bytes(&allocator, new_solution).ok()?;
-        Some(CoinSpend::new(
-            new_coin,
-            coin_spend.puzzle_reveal.clone(),
-            new_solution_bytes.into(),
-        ))
-    }
-
     pub fn get_mempool_item_by_tx_id(&self, tx_id: Bytes32) -> GetMempoolItemResponse {
         GetMempoolItemResponse {
             mempool_item: self
@@ -890,459 +715,6 @@ impl FullNodeSimulator {
         records
     }
 
-    fn validate_bundle(
-        &self,
-        spend_bundle: SpendBundle,
-    ) -> Result<ValidatedBundle, SimulatorError> {
-        if spend_bundle.coin_spends.is_empty() {
-            return Err(SimulatorError::Validation(ErrorCode::InvalidSpendBundle));
-        }
-
-        let constants = default_constants(SIMULATOR_GENESIS_CHALLENGE, SIMULATOR_GENESIS_CHALLENGE);
-        let conds = validate_clvm_and_signature(
-            &spend_bundle,
-            11_000_000_000 / 2,
-            &constants,
-            ENABLE_KECCAK_OPS_OUTSIDE_GUARD | COMPUTE_FINGERPRINT,
-        )
-        .map_err(SimulatorError::Validation)?;
-
-        if self.height < conds.height_absolute {
-            return Err(SimulatorError::Validation(
-                ErrorCode::AssertHeightAbsoluteFailed,
-            ));
-        }
-        if self.next_timestamp < conds.seconds_absolute {
-            return Err(SimulatorError::Validation(
-                ErrorCode::AssertSecondsAbsoluteFailed,
-            ));
-        }
-        if let Some(height) = conds.before_height_absolute
-            && height < self.height
-        {
-            return Err(SimulatorError::Validation(
-                ErrorCode::AssertBeforeHeightAbsoluteFailed,
-            ));
-        }
-        if let Some(seconds) = conds.before_seconds_absolute
-            && seconds < self.next_timestamp
-        {
-            return Err(SimulatorError::Validation(
-                ErrorCode::AssertBeforeSecondsAbsoluteFailed,
-            ));
-        }
-
-        let bundle_puzzle_hashes = spend_bundle
-            .coin_spends
-            .iter()
-            .map(|spend| spend.coin.puzzle_hash)
-            .collect::<HashSet<_>>();
-        let condition_puzzle_hashes = conds
-            .spends
-            .iter()
-            .map(|spend| spend.puzzle_hash)
-            .collect::<HashSet<_>>();
-        if bundle_puzzle_hashes != condition_puzzle_hashes {
-            return Err(SimulatorError::Validation(ErrorCode::InvalidSpendBundle));
-        }
-
-        let bundle_coin_spends = spend_bundle
-            .coin_spends
-            .iter()
-            .map(|spend| (spend.coin.coin_id(), spend.clone()))
-            .collect::<IndexMap<_, _>>();
-
-        let mut removals = IndexSet::new();
-        let mut additions = IndexMap::new();
-        let mut spends = IndexMap::new();
-
-        for spend in &conds.spends {
-            let coin_id = spend.coin_id;
-            let mut spend_additions = Vec::new();
-
-            for (puzzle_hash, amount, hint) in &spend.create_coin {
-                let coin = Coin::new(coin_id, *puzzle_hash, *amount);
-                let parsed_hint = hint
-                    .as_ref()
-                    .filter(|bytes| bytes.len() == 32)
-                    .and_then(|bytes| Bytes32::try_from(bytes.as_ref()).ok());
-                spend_additions.push((coin, parsed_hint));
-                additions.insert(coin.coin_id(), (coin, parsed_hint));
-            }
-
-            let Some(coin_spend) = bundle_coin_spends.get(&coin_id).cloned() else {
-                return Err(SimulatorError::Validation(ErrorCode::InvalidSpendBundle));
-            };
-
-            let fingerprint = if (spend.flags & ELIGIBLE_FOR_DEDUP) != 0 {
-                Bytes32::try_from(spend.fingerprint.as_ref()).ok()
-            } else {
-                None
-            };
-
-            spends.insert(
-                coin_id,
-                ValidatedSpend {
-                    coin_spend,
-                    flags: spend.flags,
-                    fingerprint,
-                    additions: spend_additions,
-                },
-            );
-        }
-
-        for spend in &conds.spends {
-            let coin_id = spend.coin_id;
-            if !removals.insert(coin_id) {
-                return Err(SimulatorError::Validation(ErrorCode::DoubleSpend));
-            }
-
-            if let Some(record) = self.coins.get(&coin_id) {
-                if record.spent_block_index.is_some() {
-                    return Err(SimulatorError::Validation(ErrorCode::DoubleSpend));
-                }
-
-                self.validate_relative_conditions(spend, record)?;
-            } else if additions.contains_key(&coin_id) {
-                let coin = additions
-                    .get(&coin_id)
-                    .map(|(coin, _)| *coin)
-                    .unwrap_or_else(|| {
-                        Coin::new(spend.parent_id, spend.puzzle_hash, spend.coin_amount)
-                    });
-                let ephemeral_coin_record = SimCoinRecord {
-                    coin,
-                    coinbase: false,
-                    confirmed_block_index: self.height,
-                    spent_block_index: None,
-                    timestamp: self.next_timestamp,
-                };
-                self.validate_relative_conditions(spend, &ephemeral_coin_record)?;
-            } else if let Some(coin) = self.mempool_addition_coin(coin_id) {
-                let ephemeral_coin_record = SimCoinRecord {
-                    coin,
-                    coinbase: false,
-                    confirmed_block_index: self.height,
-                    spent_block_index: None,
-                    timestamp: self.next_timestamp,
-                };
-                self.validate_relative_conditions(spend, &ephemeral_coin_record)?;
-            } else {
-                return Err(SimulatorError::Validation(ErrorCode::UnknownUnspent));
-            }
-        }
-
-        let fee = conds
-            .removal_amount
-            .checked_sub(conds.addition_amount)
-            .unwrap_or_default()
-            .try_into()
-            .unwrap_or(u64::MAX);
-        if fee < conds.reserve_fee {
-            return Err(SimulatorError::Validation(
-                ErrorCode::ReserveFeeConditionFailed,
-            ));
-        }
-
-        Ok(ValidatedBundle {
-            spend_bundle,
-            removals: removals.into_iter().collect(),
-            additions: additions.into_values().collect(),
-            spends,
-            cost: conds.cost,
-            fee,
-        })
-    }
-
-    fn mempool_addition_coin(&self, coin_id: Bytes32) -> Option<Coin> {
-        self.mempool
-            .values()
-            .flat_map(|item| item.additions.iter().map(|(coin, _)| *coin))
-            .find(|coin| coin.coin_id() == coin_id)
-    }
-
-    fn conflicting_mempool_tx_ids(&self, validated: &ValidatedBundle) -> Vec<Bytes32> {
-        self.mempool
-            .iter()
-            .filter(|(_, item)| self.has_non_dedup_overlap(validated, item))
-            .map(|(tx_id, _)| *tx_id)
-            .collect()
-    }
-
-    fn has_non_dedup_overlap(&self, lhs: &ValidatedBundle, rhs: &ValidatedBundle) -> bool {
-        lhs.removals.iter().any(|coin_id| {
-            rhs.removals.contains(coin_id) && !Self::removal_is_dedup_compatible(lhs, rhs, *coin_id)
-        })
-    }
-
-    fn removal_is_dedup_compatible(
-        lhs: &ValidatedBundle,
-        rhs: &ValidatedBundle,
-        coin_id: Bytes32,
-    ) -> bool {
-        let Some(lhs_spend) = lhs.spends.get(&coin_id) else {
-            return false;
-        };
-        let Some(rhs_spend) = rhs.spends.get(&coin_id) else {
-            return false;
-        };
-        Self::spends_are_dedup_compatible(lhs_spend, rhs_spend)
-    }
-
-    fn spends_are_dedup_compatible(lhs: &ValidatedSpend, rhs: &ValidatedSpend) -> bool {
-        (lhs.flags & ELIGIBLE_FOR_DEDUP) != 0
-            && (rhs.flags & ELIGIBLE_FOR_DEDUP) != 0
-            && lhs.fingerprint.is_some()
-            && lhs.fingerprint == rhs.fingerprint
-    }
-
-    fn validate_relative_conditions(
-        &self,
-        spend: &chia_consensus::owned_conditions::OwnedSpendConditions,
-        record: &SimCoinRecord,
-    ) -> Result<(), SimulatorError> {
-        if let Some(relative_height) = spend.height_relative
-            && self.height < record.confirmed_block_index + relative_height
-        {
-            return Err(SimulatorError::Validation(
-                ErrorCode::AssertHeightRelativeFailed,
-            ));
-        }
-        if let Some(relative_seconds) = spend.seconds_relative
-            && self.next_timestamp < record.timestamp + relative_seconds
-        {
-            return Err(SimulatorError::Validation(
-                ErrorCode::AssertSecondsRelativeFailed,
-            ));
-        }
-        if let Some(relative_height) = spend.before_height_relative
-            && record.confirmed_block_index + relative_height < self.height
-        {
-            return Err(SimulatorError::Validation(
-                ErrorCode::AssertBeforeHeightRelativeFailed,
-            ));
-        }
-        if let Some(relative_seconds) = spend.before_seconds_relative
-            && record.timestamp + relative_seconds < self.next_timestamp
-        {
-            return Err(SimulatorError::Validation(
-                ErrorCode::AssertBeforeSecondsRelativeFailed,
-            ));
-        }
-        Ok(())
-    }
-
-    fn create_block_from_mempool(&mut self) -> BlockRecord {
-        let previous_header_hash = self.header_hash();
-        let height = self.height + 1;
-        let timestamp = self.next_timestamp;
-        let header_hash = self.random_hash();
-
-        let mut included_tx_ids = Vec::new();
-        let mut included = Vec::new();
-        let mut included_spends_by_coin = IndexMap::<Bytes32, ValidatedSpend>::new();
-        for (tx_id, item) in self.mempool.clone() {
-            let Ok(validated) = self.validate_bundle(item.spend_bundle.clone()) else {
-                continue;
-            };
-            let has_conflict = validated.removals.iter().any(|coin_id| {
-                let Some(existing_spend) = included_spends_by_coin.get(coin_id) else {
-                    return false;
-                };
-                let Some(new_spend) = validated.spends.get(coin_id) else {
-                    return true;
-                };
-                !Self::spends_are_dedup_compatible(existing_spend, new_spend)
-            });
-            if has_conflict {
-                continue;
-            }
-            for coin_id in &validated.removals {
-                let Some(spend) = validated.spends.get(coin_id) else {
-                    continue;
-                };
-                included_spends_by_coin
-                    .entry(*coin_id)
-                    .or_insert_with(|| spend.clone());
-            }
-            included_tx_ids.push(tx_id);
-            included.push(validated);
-        }
-
-        for tx_id in included_tx_ids {
-            self.mempool.swap_remove(&tx_id);
-        }
-
-        let mut additions = Vec::new();
-        let mut removals = Vec::new();
-        let mut spends = Vec::new();
-        let mut transactions = Vec::new();
-        let mut previous_coin_records = Vec::new();
-        let mut added_hints = Vec::new();
-        let mut fees = 0_u64;
-        let mut applied_removals = IndexSet::new();
-        let mut applied_additions = IndexSet::new();
-        let mut applied_spends = IndexSet::new();
-        let reward_coin = Self::reward_coin(
-            header_hash,
-            height,
-            0,
-            self.farming_puzzle_hash,
-            BLOCK_REWARD_AMOUNT,
-        );
-        let reward_coin_id = reward_coin.coin_id();
-        self.insert_coin_record(reward_coin, true, height, timestamp);
-        additions.push(reward_coin_id);
-
-        for item in included {
-            fees = fees.saturating_add(item.fee);
-            transactions.push(item.spend_bundle);
-            let ephemeral_removals = item
-                .additions
-                .iter()
-                .map(|(coin, _)| coin.coin_id())
-                .collect::<IndexSet<_>>();
-            let mut pending_ephemeral_removals = Vec::new();
-
-            for coin_id in item.removals {
-                if let Some(record) = self.coins.get_mut(&coin_id) {
-                    if !applied_removals.insert(coin_id) {
-                        continue;
-                    }
-                    previous_coin_records.push((coin_id, *record));
-                    record.spent_block_index = Some(height);
-                    removals.push(coin_id);
-                } else if ephemeral_removals.contains(&coin_id) {
-                    pending_ephemeral_removals.push(coin_id);
-                }
-            }
-
-            for (coin, hint) in item.additions {
-                let coin_id = coin.coin_id();
-                if !applied_additions.insert(coin_id) {
-                    continue;
-                }
-                self.insert_coin_record(coin, false, height, timestamp);
-                if let Some(hint) = hint {
-                    self.coin_hints.insert(coin_id, hint);
-                    added_hints.push(coin_id);
-                }
-                additions.push(coin_id);
-            }
-
-            for coin_id in pending_ephemeral_removals {
-                if !applied_removals.insert(coin_id) {
-                    continue;
-                }
-                if let Some(record) = self.coins.get_mut(&coin_id) {
-                    record.spent_block_index = Some(height);
-                    removals.push(coin_id);
-                }
-            }
-
-            for (coin_id, spend) in item.spends {
-                if applied_spends.insert(coin_id) {
-                    spends.push(spend.coin_spend);
-                }
-            }
-        }
-
-        for spend in &spends {
-            self.coin_spends.insert(spend.coin.coin_id(), spend.clone());
-        }
-
-        let record = Self::make_block_record(
-            header_hash,
-            previous_header_hash,
-            height,
-            timestamp,
-            self.header_hash_of(height.saturating_sub(1))
-                .unwrap_or_default(),
-            fees,
-            height.saturating_sub(1),
-            self.farming_puzzle_hash,
-            vec![reward_coin],
-        );
-        let block = SimBlock {
-            record: record.clone(),
-            additions: additions.clone(),
-            removals: removals.clone(),
-            spends,
-            transactions,
-            previous_coin_records,
-            added_hints,
-        };
-
-        self.blocks.insert(header_hash, block);
-        self.header_hashes.push(header_hash);
-        self.height = height;
-        self.next_timestamp = self.next_timestamp.saturating_add(1);
-
-        self.events.push(FullNodeSimulatorEvent::Block {
-            height,
-            header_hash,
-            previous_header_hash,
-            additions: self.records_for_ids(&additions),
-            removals: self.records_for_ids(&removals),
-        });
-
-        record
-    }
-
-    fn revert_canonical_blocks(&mut self, blocks: u32) -> Vec<SimBlock> {
-        let mut reverted = Vec::new();
-        for _ in 0..blocks {
-            if self.height == 0 {
-                break;
-            }
-            let Some(header_hash) = self.header_hashes.pop() else {
-                break;
-            };
-            let Some(block) = self.blocks.swap_remove(&header_hash) else {
-                break;
-            };
-
-            for coin_id in &block.additions {
-                self.coins.swap_remove(coin_id);
-                self.coin_hints.swap_remove(coin_id);
-            }
-            for coin_id in &block.added_hints {
-                self.coin_hints.swap_remove(coin_id);
-            }
-            for (coin_id, previous_record) in &block.previous_coin_records {
-                self.coins.insert(*coin_id, *previous_record);
-                self.coin_spends.swap_remove(coin_id);
-            }
-
-            self.height = self.height.saturating_sub(1);
-            self.next_timestamp = block.record.timestamp.unwrap_or(self.next_timestamp);
-            reverted.push(block);
-        }
-        reverted
-    }
-
-    fn requeue_transactions(&mut self, transactions: impl IntoIterator<Item = SpendBundle>) {
-        for spend_bundle in transactions {
-            let tx_id = spend_bundle.name();
-            if self.mempool.contains_key(&tx_id) {
-                continue;
-            }
-            if let Ok(validated) = self.validate_bundle(spend_bundle) {
-                let _ = self.insert_mempool_item(tx_id, validated);
-            }
-        }
-    }
-
-    fn prune_mempool(&mut self) {
-        let spend_bundles = self
-            .mempool
-            .values()
-            .map(|item| item.spend_bundle.clone())
-            .collect::<Vec<_>>();
-        self.mempool.clear();
-        self.requeue_transactions(spend_bundles);
-    }
-
     fn insert_coin_record(&mut self, coin: Coin, coinbase: bool, height: u32, timestamp: u64) {
         self.coins.insert(
             coin.coin_id(),
@@ -1386,6 +758,8 @@ impl FullNodeSimulator {
             coin_records: Some(records),
             error: None,
             success: true,
+            next_cursor: None,
+            truncated: None,
         }
     }
 
@@ -1493,16 +867,9 @@ impl ValidatedBundle {
 mod tests {
     use chia_bls::{SecretKey, Signature};
     use chia_protocol::{Coin, CoinSpend, SpendBundle};
-    use chia_puzzle_types::{
-        DeriveSynthetic, LineageProof, Proof,
-        singleton::{SingletonArgs, SingletonSolution},
-        standard::StandardArgs,
-    };
-    use chia_sdk_types::Mod;
+    use chia_puzzle_types::{DeriveSynthetic, standard::StandardArgs};
     use chia_sdk_types::conditions::{CreateCoin, Memos};
-    use clvm_traits::ToClvm;
-    use clvm_utils::CurriedProgram;
-    use clvmr::{Allocator, NodePtr, serde::node_from_bytes, serde::node_to_bytes};
+    use clvmr::NodePtr;
 
     use crate::{FullNodeSimulatorEvent, to_program, to_puzzle};
 
@@ -1521,53 +888,6 @@ mod tests {
                 to_program([CreateCoin::<NodePtr>::new(puzzle_hash, amount, Memos::None)])?,
             )],
             Signature::default(),
-        ))
-    }
-
-    fn singleton_spend_to_child(
-        coin: Coin,
-        launcher_id: Bytes32,
-        inner_puzzle_reveal: chia_protocol::Program,
-        lineage_proof: LineageProof,
-        child_puzzle_hash: Bytes32,
-        child_amount: u64,
-        hint: Option<Bytes32>,
-    ) -> anyhow::Result<CoinSpend> {
-        let mut allocator = Allocator::new_limited(500_000_000);
-        let memos = if let Some(hint) = hint {
-            let hint_atom = allocator.new_atom(hint.as_ref())?;
-            let memo_list = allocator.new_pair(hint_atom, NodePtr::NIL)?;
-            Memos::Some(memo_list)
-        } else {
-            Memos::None
-        };
-        let inner_solution = [CreateCoin::<NodePtr>::new(
-            child_puzzle_hash,
-            child_amount,
-            memos,
-        )]
-        .to_clvm(&mut allocator)?;
-        let singleton_mod = node_from_bytes(
-            &mut allocator,
-            SingletonArgs::<NodePtr>::mod_reveal().as_ref(),
-        )?;
-        let inner_puzzle = node_from_bytes(&mut allocator, inner_puzzle_reveal.as_slice())?;
-        let singleton_puzzle = CurriedProgram {
-            program: singleton_mod,
-            args: SingletonArgs::new(launcher_id, inner_puzzle),
-        }
-        .to_clvm(&mut allocator)?;
-        let singleton_solution = SingletonSolution {
-            lineage_proof: Proof::Lineage(lineage_proof),
-            amount: coin.amount,
-            inner_solution,
-        }
-        .to_clvm(&mut allocator)?;
-
-        Ok(CoinSpend::new(
-            coin,
-            node_to_bytes(&allocator, singleton_puzzle)?.into(),
-            node_to_bytes(&allocator, singleton_solution)?.into(),
         ))
     }
 
@@ -1805,312 +1125,6 @@ mod tests {
                 .mempool_size,
             0
         );
-
-        Ok(())
-    }
-
-    #[test]
-    fn push_tx_rejects_mempool_conflict() -> anyhow::Result<()> {
-        let mut sim = FullNodeSimulator::new();
-        sim.set_autofarm(false);
-        let (puzzle_hash, puzzle_reveal) = to_puzzle(1)?;
-        let coin = sim.new_coin(puzzle_hash, 100);
-        let first = spend_to_child(coin, puzzle_reveal.clone(), puzzle_hash, 98)?;
-        let conflicting = spend_to_child(coin, puzzle_reveal, puzzle_hash, 99)?;
-
-        assert!(sim.push_tx(first).success);
-        let response = sim.push_tx(conflicting);
-
-        assert!(!response.success);
-        assert_eq!(
-            response.error.as_deref(),
-            Some("Validation error: MempoolConflict")
-        );
-        assert_eq!(
-            sim.get_blockchain_state()
-                .blockchain_state
-                .unwrap()
-                .mempool_size,
-            1
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn push_tx_replaces_mempool_conflict_with_higher_fee_superset() -> anyhow::Result<()> {
-        let mut sim = FullNodeSimulator::new();
-        sim.set_autofarm(false);
-        let (puzzle_hash, puzzle_reveal) = to_puzzle(1)?;
-        let coin = sim.new_coin(puzzle_hash, 100);
-        let first = spend_to_child(coin, puzzle_reveal.clone(), puzzle_hash, 99)?;
-        let replacement = spend_to_child(coin, puzzle_reveal, puzzle_hash, 98)?;
-        let first_tx_id = first.name();
-        let replacement_tx_id = replacement.name();
-
-        assert!(sim.push_tx(first).success);
-        assert!(sim.push_tx(replacement).success);
-
-        assert!(
-            sim.get_mempool_item_by_tx_id(first_tx_id)
-                .mempool_item
-                .is_none()
-        );
-        assert!(
-            sim.get_mempool_item_by_tx_id(replacement_tx_id)
-                .mempool_item
-                .is_some()
-        );
-        assert_eq!(
-            sim.get_blockchain_state()
-                .blockchain_state
-                .unwrap()
-                .mempool_size,
-            1
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn push_tx_does_not_replace_conflict_that_is_not_a_superset() -> anyhow::Result<()> {
-        let mut sim = FullNodeSimulator::new();
-        sim.set_autofarm(false);
-        let (puzzle_hash, puzzle_reveal) = to_puzzle(1)?;
-        let coin_a = sim.new_coin(puzzle_hash, 100);
-        let coin_b = sim.new_coin(puzzle_hash, 100);
-        let first_spend_a = spend_to_child(coin_a, puzzle_reveal.clone(), puzzle_hash, 99)?;
-        let first_spend_b = spend_to_child(coin_b, puzzle_reveal.clone(), puzzle_hash, 99)?;
-        let first = SpendBundle::new(
-            vec![
-                first_spend_a.coin_spends[0].clone(),
-                first_spend_b.coin_spends[0].clone(),
-            ],
-            Signature::default(),
-        );
-        let conflicting = spend_to_child(coin_a, puzzle_reveal, puzzle_hash, 50)?;
-
-        assert!(sim.push_tx(first).success);
-        let response = sim.push_tx(conflicting);
-
-        assert!(!response.success);
-        assert_eq!(
-            response.error.as_deref(),
-            Some("Validation error: MempoolConflict")
-        );
-        assert_eq!(
-            sim.get_blockchain_state()
-                .blockchain_state
-                .unwrap()
-                .mempool_size,
-            1
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn push_tx_allows_dedup_compatible_mempool_overlap() -> anyhow::Result<()> {
-        let mut sim = FullNodeSimulator::new();
-        sim.set_autofarm(false);
-        let (puzzle_hash, puzzle_reveal) = to_puzzle(1)?;
-        let shared_coin = sim.new_coin(puzzle_hash, 100);
-        let extra_coin = sim.new_coin(puzzle_hash, 100);
-        let shared_spend = spend_to_child(shared_coin, puzzle_reveal.clone(), puzzle_hash, 100)?;
-        let extra_spend = spend_to_child(extra_coin, puzzle_reveal, puzzle_hash, 99)?;
-        let second_bundle = SpendBundle::new(
-            vec![
-                shared_spend.coin_spends[0].clone(),
-                extra_spend.coin_spends[0].clone(),
-            ],
-            Signature::default(),
-        );
-
-        assert!(sim.push_tx(shared_spend).success);
-        assert!(sim.push_tx(second_bundle).success);
-        assert_eq!(
-            sim.get_blockchain_state()
-                .blockchain_state
-                .unwrap()
-                .mempool_size,
-            2
-        );
-
-        sim.farm_block(1);
-        let spends = sim
-            .get_block_spends(sim.header_hash())
-            .block_spends
-            .unwrap();
-        assert_eq!(spends.len(), 2);
-        assert_eq!(
-            spends
-                .iter()
-                .filter(|spend| spend.coin.coin_id() == shared_coin.coin_id())
-                .count(),
-            1
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn try_fast_forward_rewrites_singleton_spend_against_mempool_item() -> anyhow::Result<()> {
-        let mut sim = FullNodeSimulator::new();
-        let (inner_puzzle_hash, inner_puzzle_reveal) = to_puzzle(1)?;
-        let launcher_id: Bytes32 = [7; 32].into();
-        let singleton_puzzle_hash: Bytes32 =
-            SingletonArgs::curry_tree_hash(launcher_id, inner_puzzle_hash.into()).into();
-        let parent_coin = Coin::new([9; 32].into(), singleton_puzzle_hash, 101);
-        let singleton_coin = Coin::new(parent_coin.coin_id(), singleton_puzzle_hash, 101);
-        let lineage_proof = LineageProof {
-            parent_parent_coin_info: parent_coin.parent_coin_info,
-            parent_inner_puzzle_hash: inner_puzzle_hash,
-            parent_amount: parent_coin.amount,
-        };
-        sim.insert_coin(singleton_coin);
-
-        let first_singleton_spend = singleton_spend_to_child(
-            singleton_coin,
-            launcher_id,
-            inner_puzzle_reveal.clone(),
-            lineage_proof,
-            singleton_puzzle_hash,
-            singleton_coin.amount,
-            None,
-        )?;
-        let fast_forward_hint: Bytes32 = [8; 32].into();
-        let second_singleton_spend = singleton_spend_to_child(
-            singleton_coin,
-            launcher_id,
-            inner_puzzle_reveal.clone(),
-            lineage_proof,
-            singleton_puzzle_hash,
-            singleton_coin.amount,
-            Some(fast_forward_hint),
-        )?;
-        let child_coin = Coin::new(
-            singleton_coin.coin_id(),
-            singleton_coin.puzzle_hash,
-            singleton_coin.amount,
-        );
-        let first_tx = SpendBundle::new(vec![first_singleton_spend.clone()], Signature::default());
-        sim.mempool.insert(
-            first_tx.name(),
-            ValidatedBundle {
-                spend_bundle: first_tx,
-                removals: vec![singleton_coin.coin_id()],
-                additions: vec![(child_coin, None)],
-                spends: IndexMap::from([(
-                    singleton_coin.coin_id(),
-                    ValidatedSpend {
-                        coin_spend: first_singleton_spend,
-                        flags: ELIGIBLE_FOR_FF,
-                        fingerprint: None,
-                        additions: vec![(child_coin, None)],
-                    },
-                )]),
-                cost: 0,
-                fee: 0,
-            },
-        );
-
-        let candidate_bundle =
-            SpendBundle::new(vec![second_singleton_spend.clone()], Signature::default());
-        let rewritten = sim
-            .try_fast_forward_bundle(&ValidatedBundle {
-                spend_bundle: candidate_bundle.clone(),
-                removals: vec![singleton_coin.coin_id()],
-                additions: Vec::new(),
-                spends: IndexMap::from([(
-                    singleton_coin.coin_id(),
-                    ValidatedSpend {
-                        coin_spend: second_singleton_spend,
-                        flags: ELIGIBLE_FOR_FF,
-                        fingerprint: None,
-                        additions: Vec::new(),
-                    },
-                )]),
-                cost: 0,
-                fee: 0,
-            })
-            .expect("singleton spend should be fast-forwarded");
-        assert!(
-            rewritten
-                .coin_spends
-                .iter()
-                .any(|spend| spend.coin.coin_id() == child_coin.coin_id())
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn push_tx_fast_forwards_already_settled_singleton_spend() -> anyhow::Result<()> {
-        let mut sim = FullNodeSimulator::new();
-        let (inner_puzzle_hash, inner_puzzle_reveal) = to_puzzle(1)?;
-        let launcher_id: Bytes32 = [11; 32].into();
-        let singleton_puzzle_hash: Bytes32 =
-            SingletonArgs::curry_tree_hash(launcher_id, inner_puzzle_hash.into()).into();
-        let parent_coin = Coin::new([13; 32].into(), singleton_puzzle_hash, 101);
-        let singleton_coin = Coin::new(parent_coin.coin_id(), singleton_puzzle_hash, 101);
-        let lineage_proof = LineageProof {
-            parent_parent_coin_info: parent_coin.parent_coin_info,
-            parent_inner_puzzle_hash: inner_puzzle_hash,
-            parent_amount: parent_coin.amount,
-        };
-        sim.insert_coin(singleton_coin);
-        let child_coin = Coin::new(
-            singleton_coin.coin_id(),
-            singleton_coin.puzzle_hash,
-            singleton_coin.amount,
-        );
-        sim.insert_coin(child_coin);
-        sim.coins
-            .get_mut(&singleton_coin.coin_id())
-            .unwrap()
-            .spent_block_index = Some(2);
-
-        let stale_singleton_spend = singleton_spend_to_child(
-            singleton_coin,
-            launcher_id,
-            inner_puzzle_reveal.clone(),
-            lineage_proof,
-            singleton_puzzle_hash,
-            singleton_coin.amount,
-            Some([14; 32].into()),
-        )?;
-        let stale_bundle = SpendBundle::new(vec![stale_singleton_spend], Signature::default());
-        assert!(
-            FullNodeSimulator::fast_forward_coin_spend(
-                &stale_bundle.coin_spends[0],
-                child_coin,
-                singleton_coin,
-            )
-            .is_some()
-        );
-        let maybe_rewritten = sim.try_fast_forward_settled_bundle(&stale_bundle);
-        assert!(maybe_rewritten.is_some());
-        let rewritten = maybe_rewritten.unwrap();
-        assert_eq!(
-            rewritten.coin_spends[0].coin.coin_id(),
-            child_coin.coin_id()
-        );
-
-        let response = sim.push_tx(stale_bundle);
-        assert!(response.success, "{response:?}");
-
-        let child_record = sim
-            .get_coin_record_by_name(child_coin.coin_id())
-            .coin_record
-            .unwrap();
-        assert!(child_record.spent);
-
-        let last_spends = sim
-            .get_block_spends(sim.header_hash())
-            .block_spends
-            .unwrap();
-        assert_eq!(last_spends.len(), 1);
-        assert_eq!(last_spends[0].coin.coin_id(), child_coin.coin_id());
 
         Ok(())
     }
